Comprehensive Chech Run ID Resolution
This was surprisingly difficult.  This is pretty reliable, using low-cost methods (filtering on Name comparison, Annotation count [we add one, so it should have atleast one]) and then searching for a unique string if the count is still higher than 1.

I'm pretty happy about this, I think this will stand the test of time until GitHub get's their act together...