Update README.md
I recently came across a bug where specifying `go-version: '1.20'` used Go 1.2, not 1.20. It was not clear from the readme that to avoid this, the Go version should be written inside single quotes (as per this comment: https://github.com/actions/setup-go/issues/326#issuecomment-1415719692) but the single quotes are missing from the example in this repository.  This PR fixes this.